Extract `MatchChessBoardState` from `MatchGameScreenState`
