+++ mobile/src/androidTest/java/ch/epfl/sdp/mobile/test/state/StatefulArScreenTest.kt
@@ -7,7 +7,6 @@ import androidx.test.rule.GrantPermissionRule
 import androidx.test.rule.GrantPermissionRule.grant
 import ch.epfl.sdp.mobile.application.ChessDocument
 import ch.epfl.sdp.mobile.application.ProfileDocument
-import ch.epfl.sdp.mobile.application.authentication.AuthenticatedUser
 import ch.epfl.sdp.mobile.application.authentication.AuthenticationFacade
 import ch.epfl.sdp.mobile.application.chess.ChessFacade
 import ch.epfl.sdp.mobile.application.social.SocialFacade
@@ -17,8 +16,6 @@ import ch.epfl.sdp.mobile.state.StatefulArScreen
 import ch.epfl.sdp.mobile.test.infrastructure.persistence.auth.emptyAuth
 import ch.epfl.sdp.mobile.test.infrastructure.persistence.store.buildStore
 import ch.epfl.sdp.mobile.test.infrastructure.persistence.store.document
-import io.mockk.every
-import io.mockk.mockk
 import org.junit.Rule
 import org.junit.Test
 
@@ -44,12 +41,9 @@ class StatefulArScreenTest {
     val social = SocialFacade(auth, store)
     val chess = ChessFacade(auth, store)
 
-    val user1 = mockk<AuthenticatedUser>()
-    every { user1.uid } returns "userId1"
-
     val strings =
         rule.setContentWithLocalizedStrings {
-          ProvideFacades(authApi, social, chess) { StatefulArScreen(user1, "gameId") }
+          ProvideFacades(authApi, social, chess) { StatefulArScreen("gameId") }
         }
 
     rule.onNodeWithContentDescription(strings.arContentDescription).assertExists()
